+++ integration_tests/juniper_tests/fail/object/impl_no_underscore_arg.rs
@@ -1,0 +1,11 @@
+struct Object {}
+
+#[juniper::graphql_object]
+impl Object {
+    #[graphql(arguments(arg(name = "__arg")))]
+    fn test(&self, arg: String) -> String {
+        arg
+    }
+}
+
+fn main() {}