+++ integration_tests/juniper_tests/src/codegen/mod.rs
@@ -5,6 +5,7 @@ mod derive_object_with_raw_idents;
 mod derive_union;
 mod impl_scalar;
 mod impl_union;
+mod impl_object;
 mod scalar_value_transparent;
 
 use std::{