+++ juniper_codegen/src/derive_input_object.rs
@@ -68,7 +68,7 @@ pub fn impl_input_object(
 
             if name.starts_with("__") {
                 error.no_double_underscore(if let Some(name) = field_attrs.name {
-                    name.span()
+                    name.span_ident()
                 } else {
                     name.span()
                 });
@@ -123,7 +123,7 @@ pub fn impl_input_object(
 
     if name.starts_with("__") && !is_internal {
         error.no_double_underscore(if let Some(name) = attrs.name {
-            name.span()
+            name.span_ident()
         } else {
             ident.span()
         });