+++ juniper_codegen/src/derive_union.rs
@@ -101,7 +101,7 @@ pub fn build_derive_union(
 
             if name.starts_with("__") {
                 error.no_double_underscore(if let Some(name) = field_attrs.name {
-                    name.span()
+                    name.span_ident()
                 } else {
                     variant_name.span()
                 });
@@ -137,7 +137,7 @@ pub fn build_derive_union(
 
     if name.starts_with("__") && !is_internal {
         error.no_double_underscore(if let Some(name) = attrs.name {
-            name.span()
+            name.span_ident()
         } else {
             ident.span()
         });